pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.189.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.2",
    "tweepy>=4.14.0",
    "watchdog>=6.0.0",
]
